requirements.txt
alembic==0.8.10

apache-airflow==1.8.1

certifi==2017.4.17

Flask==0.11.1

Flask-Admin==1.4.1

Flask-Cache==0.13.1

Flask-Login==0.2.11

flask-swagger==0.2.13

Flask-WTF==0.12

gitdb2==2.0.2

GitPython==2.1.5

gunicorn==19.3.0

lxml==3.8.0

numpy==1.13.1

pandas==0.20.3

psutil==4.4.2

psycopg2==2.7.3

Pygments==2.2.0

python-daemon==2.1.2

python-dateutil==2.6.1

python-editor==1.0.3

python-nvd3==0.14.2

python-slugify==1.1.4

pytz==2017.2

PyYAML==3.12

requests==2.18.1

SQLAlchemy==1.1.11

tabulate==0.7.7

thrift==0.9.3

Unidecode==0.4.21

urllib3==1.21.1

line-bot-sdk
bs4
pymongo
datetime
scipy
msgpack
